requirements.txt
python-dotenv
discord.py

flask
py-cord
